Add settings view for tip default amount
Ensures that when a user navigates to the settings view, the previously
selected default tip % is pre-selected.

Adds an observer to UserDefault so that the tip % in the home screen is
only changed when the user changes the default, not every time
viewWillAppear is called. This is because if a user visits the settings
page but doesn’t change the default, we don’t want the tip % in the
home screen to change to the default.